test(scrna): merge tests for Seurat processes to avoid repeatedly load pbmc3k dataset
- Deleted run.env and test.py files from SeuratClusterStats, SeuratFilter, SeuratMap2Ref, SeuratPreparing, SeuratSubset test directories.
- Added new test.py file for Seurat with comprehensive pipeline setup including data preparation, clustering, annotation, and statistical analysis.
- Introduced PrepareMetafile class for generating input metafiles for SeuratPreparing.
- Updated environment configurations for various classes to enhance functionality and ensure compatibility with the latest Seurat features.
- Added support for SCCatch cell type annotation and integrated it into the SeuratClusterStats workflow.
