--- tests/test_scrna/ScCATCH/run.env
@@ -1,1 +1,0 @@
-ENVNAME="biopipen"
