--- tests/test_scrna/ScFGSEA/run.env
@@ -1,1 +1,0 @@
-ENVNAME="biopipen"
