--- tests/test_scrna/SeuratClusterStats/run.env
@@ -1,1 +1,0 @@
-ENVNAME="biopipen"
