--- tests/test_scrna/SeuratFilter/run.env
@@ -1,1 +1,0 @@
-ENVNAME="biopipen"
