--- tests/test_scrna/SeuratMap2Ref/run.env
@@ -1,2 +1,0 @@
-#ENVNAME="biopipen"
-LOCAL_ONLY="true"
