--- tests/test_scrna/SeuratPreparing/run.env
@@ -1,1 +1,0 @@
-ENVNAME="biopipen"
