--- tests/test_scrna/SeuratSubset/run.env
@@ -1,1 +1,0 @@
-ENVNAME="biopipen"
